Refactor errcode.
